pyproject.toml
[project]
name = "auto-rp"
version = "0.1.0"
description = "Automated roleplay generator"
readme = "docs/README.md"
requires-python = ">=3.8"
dependencies = [
    "click",
    "rich",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
auto-rp = "main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
